Python3 update
reconfigured to Python3
Also reorganised __init so that all is imported.
So Draft, Drawstyle, Color factored out into own files
